Refactor templates into their own files
Editing templates that are hard-coded strings in a Go source file is a
pain in the ass. Instead of having to do that, it's better to edit the
templates as HTML, and then have them injected somehow.

The way this works is by doing it similarly to how JS workflows do it
with Gulp et all. The templates are edited in their own source files,
and then a Rake task generates the actual Go source file with the
templates included in them.

The need to introduce multiple toolchains meant needing to create an
Envctl environment for it, which also meant vendoring all the go
dependencies, which is why `dep` was added.